move test data around, fix test issues, spark3 test config
